+++ src/test/com/epam/esm/data_provider/StaticDataProvider.java
@@ -3,6 +3,7 @@ package com.epam.esm.data_provider;
 import com.epam.esm.model.entity.GiftCertificate;
 import com.epam.esm.model.entity.Tag;
 
+import java.math.BigDecimal;
 import java.util.Collections;
 import java.util.List;
 
@@ -19,7 +20,7 @@ public class StaticDataProvider {
     public static final List<Tag> TAG_LIST;
 
     static {
-        TAG = new Tag(1, "Вязание");
+        TAG = new Tag(1L, "Вязание");
         ADDING_TAG = new Tag();
         ADDING_TAG.setName("Вязание");
         UPDATED_TAG = new Tag(TAG.getId(), "Вышивание");
@@ -27,14 +28,14 @@ public class StaticDataProvider {
         ADDING_CERTIFICATE.setName("Курсы вышивания");
         ADDING_CERTIFICATE.setDescription("Лучшие курсы вышивания от компании Vyazhem Vmeste");
         ADDING_CERTIFICATE.setDuration(30);
-        ADDING_CERTIFICATE.setPrice(100);
+        ADDING_CERTIFICATE.setPrice(BigDecimal.valueOf(100));
         ADDING_CERTIFICATE.setTags(Collections.singletonList(TAG));
         GIFT_CERTIFICATE = new GiftCertificate(ADDING_CERTIFICATE);
         GIFT_CERTIFICATE.setId(1L);
         GIFT_CERTIFICATE.setCreateDate(ISO_DATE);
         GIFT_CERTIFICATE.setLastUpdateDate(ISO_DATE);
         UPDATED_GIFT_CERTIFICATE = new GiftCertificate(GIFT_CERTIFICATE);
-        UPDATED_GIFT_CERTIFICATE.setPrice(60);
+        UPDATED_GIFT_CERTIFICATE.setPrice(BigDecimal.valueOf(60));
         GIFT_CERTIFICATE_LIST = Collections.singletonList(GIFT_CERTIFICATE);
         TAG_LIST = Collections.singletonList(TAG);
     }
